Fixed substitution code
